pydesktopbrowerrecorder/requirements.txt
imageio-ffmpeg
numpy~=1.22.2
imageio==2.15.0
PyAutoGUI~=0.9.53
selenium~=4.1.0
Pillow~=9.0.1